requirements.txt
transformers==4.57.1
accelerate==1.10.1
peft==0.17.1
pillow==11.3.0
runpod==1.6.2
requests==2.32.5
#unsloth @ git+https://github.com/unslothai/unsloth.git@0a69adb79ab9c405f504b19d2969c875a41ec430
unsloth==2025.11.3
unsloth_zoo==2025.11.4
torch==2.8.0
torchao==0.13.0
torchaudio==2.8.0
torchvision==0.23.0
bitsandbytes==0.48.1
triton==3.4.0
#transformers @ git+https://github.com/huggingface/transformers.git
ultralytics==8.3.227
ultralytics-thop==2.0.18
langchain>=0.1.0
langchain-community
langchain-classic
langchain-core
langchain-huggingface
duckduckgo-search